requirements.txt
# NeboCollections AI — Python dependencies
# Generated for Render deployment.
# Pin exact versions to guarantee reproducible builds.
# To update: pip install -U <package> then update the version here.

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# HTTP client (async)
httpx==0.27.2

# Data validation
pydantic==2.9.2

# Environment variables
python-dotenv==1.0.1

# AI — Gemini (Free/Plus tier)
google-generativeai==0.8.3

# AI — Anthropic / Claude (Pro tier, Phase 4)
# Uncomment when wiring up claude-sonnet-4-5
# anthropic==0.40.0
groq==0.11.0
